+++ src/tree.js
@@ -1,6 +1,7 @@
 import _ from 'lodash/fp'
 import { push } from './array'
 import { findIndexed } from './conversion'
+import { dotEncoder, slashEncoder } from './array'
 
 export let isTraversable = x => _.isArray(x) || _.isPlainObject(x)
 export let traverse = x => isTraversable(x) && !_.isEmpty(x) && x
@@ -61,6 +62,22 @@ export let keyTreeByWith = (next = traverse) =>
     )(x)
   )
 
+// Flat Tree
+export let treeKeys = (x, i, xs, is) => [i, ...is]
+export let treeValues = (x, i, xs, is) => [x, ...xs]
+export let treePath = (build = treeKeys, encoder = dotEncoder) => (...args) =>
+  (encoder.encode || encoder)(build(...args).reverse())
+export let propTreePath = prop =>
+  treePath(_.flow(treeValues, _.map(prop)), slashEncoder)
+
+export let flattenTree = (next = traverse) => (buildPath = treePath()) =>
+  reduceTree(next)(
+    (result, node, ...x) => _.set([buildPath(node, ...x)], node, result),
+    {}
+  )
+
+export let flatLeaves = (next = traverse) => _.reject(next)
+
 export let tree = (next = traverse, buildIteratee = _.identity) => ({
   walk: walk(next),
   transform: transformTree(next),
@@ -70,4 +87,7 @@ export let tree = (next = traverse, buildIteratee = _.identity) => ({
   leaves: leaves(next),
   lookup: treeLookup(next, buildIteratee),
   keyByWith: keyTreeByWith(next),
+  traverse: next,
+  flatten: flattenTree(next),
+  flatLeaves: flatLeaves(next)
 })